airflow/requirements.txt
fastembed==0.2.6
weaviate-client==4.4.1
apache-airflow-providers-weaviate==1.1.0
numpy==1.26.4
pandas==2.0.3
requests==2.31.0